Move styled components for gallery out of file.
Add new gallery borders.
