+++ src/partials/gallery/preview.js
@@ -100,4 +100,5 @@ export default Preview;
 const PreviewContainer = styled.div`
   background-color: #ffffff;
   box-shadow: -12px 12px 0px 0px rgba(188, 82, 64, 1);
+  border: 0.5px solid #f2f2f2;
 `;
